Extract reusable POSTotalsSectionView from order details
Create a shared totals breakdown card component that displays subtotal,
discount, tax, total, paid amount with payment method, and optional
refund rows with net payment. Adopt it in POSOrderDetailsView, removing
the inline totals section and its associated localization strings.
